+++ packages/lib/package/plugins/index.ts
@@ -1,9 +1,5 @@
-export {
-  plugin as installToaster,
-  ToasterHandler,
-  ToasterOptions,
-  ToasterPositions,
-} from './toaster'
+export { plugin as installToaster, ToasterHandler } from './toaster'
+export type { ToasterOptions, ToasterPositions } from './toaster'
 
 export {
   plugin as installWait,
@@ -11,9 +7,5 @@ export {
   instance as waitInstance,
 } from './wait'
 
-export {
-  plugin as installAos,
-  instance as aosInstance,
-  AosHandler,
-  AosOptions,
-} from './aos'
+export { plugin as installAos, instance as aosInstance } from './aos'
+export type { AosHandler, AosOptions } from './aos'